Adds initialization logic for BTreeState
I also added a new package for serializers. I initially thought
it would be a good idea to provide TypeSerializers for the
internal and leaf page objects. However, I quickly back-tracked
on this initiative and used the TypeInformation/TypeHint utils
that flink gives out of the box.

I will return later to the serialzers package when I take a
stab at having this BTree represent records as bytes rather
than objects.